perf: code refactor

+++ pkg/customizer/customizer_td.go
@@ -140,7 +140,7 @@ func (c *Customizer) buildLowestLevelTD(
 						} else {
 							// found an exit point of the cell
 							// save this shortcut travelTime
-							exitOverlay, _ := c.graph.GetOverlayVertex(uId, uint8(exitPoint), true) // overlay vetex id of exit vertex c_1(u).
+							exitOverlay, _ := c.graph.GetOverlayVertex(uId, int(exitPoint), true) // overlay vetex id of exit vertex c_1(u).
 
 							if _, ok := overlayTravelTime[exitOverlay]; !ok {
 								overlayTravelTime[exitOverlay] = exitTTF
